reorganized user components
