+++ client/src/components/UserComponents/UserStocks/index.js
